Move exceptions to exceptions package
