--- api/src/main/java/net/quickwrite/fluent4j/exception/RecursionDepthReachedException.java
@@ -1,4 +1,0 @@
-package net.quickwrite.fluent4j.exception;
-
-public class RecursionDepthReachedException extends RuntimeException {
-}
