+++ impl/src/main/java/net/quickwrite/fluent4j/impl/parser/base/CommentSkipper.java
@@ -1,7 +1,7 @@
 package net.quickwrite.fluent4j.impl.parser.base;
 
 import net.quickwrite.fluent4j.ast.entry.FluentEntry;
-import net.quickwrite.fluent4j.container.exception.FluentExpectedException;
+import net.quickwrite.fluent4j.exception.FluentExpectedException;
 import net.quickwrite.fluent4j.iterator.ContentIterator;
 import net.quickwrite.fluent4j.parser.base.FluentElementParser;
 import net.quickwrite.fluent4j.parser.result.ParseResult;
