+++ impl/src/main/java/net/quickwrite/fluent4j/impl/parser/pattern/FluentPlaceableParser.java
@@ -3,8 +3,8 @@ package net.quickwrite.fluent4j.impl.parser.pattern;
 import net.quickwrite.fluent4j.ast.FluentPattern;
 import net.quickwrite.fluent4j.ast.placeable.FluentPlaceable;
 import net.quickwrite.fluent4j.ast.placeable.FluentSelect;
-import net.quickwrite.fluent4j.container.exception.FluentBuilderException;
-import net.quickwrite.fluent4j.container.exception.FluentExpectedException;
+import net.quickwrite.fluent4j.exception.FluentBuilderException;
+import net.quickwrite.fluent4j.exception.FluentExpectedException;
 import net.quickwrite.fluent4j.impl.ast.pattern.FluentNumberLiteral;
 import net.quickwrite.fluent4j.impl.ast.pattern.FluentSelectExpression;
 import net.quickwrite.fluent4j.impl.ast.pattern.FluentTextElement;
