+++ impl/src/main/java/net/quickwrite/fluent4j/impl/parser/pattern/placeable/FluentVariableReferenceParser.java
@@ -1,7 +1,7 @@
 package net.quickwrite.fluent4j.impl.parser.pattern.placeable;
 
 import net.quickwrite.fluent4j.ast.placeable.FluentPlaceable;
-import net.quickwrite.fluent4j.container.exception.FluentExpectedException;
+import net.quickwrite.fluent4j.exception.FluentExpectedException;
 import net.quickwrite.fluent4j.impl.ast.pattern.FluentVariableReference;
 import net.quickwrite.fluent4j.impl.util.ParserUtil;
 import net.quickwrite.fluent4j.iterator.ContentIterator;
